refactor: decompose SearchPage into hooks and sub-components
- Extract useSearchStream hook (SSE streaming, session management, phase
  tracking, offline detection, slow hint timer)
- Create ChatMessageList component (user/assistant bubbles, follow-ups,
  streaming skeleton, error/blocked/offline states, end ref)
- Create SearchInputBar component (ref-based textarea resize, Enter/Shift+Enter,
  send button with disabled states, new conversation toggle)
- Refactor SearchPage from 742-line monolith to ~185-line orchestrator
- Fix textarea to use ref-based height management instead of inline e.target
- Add error handling to SaiSearchLayout history panel (error state, retry,
  optimistic delete with rollback, cursor ref for stale closure)
