+++ components/ai-search/sai-search-layout.tsx
@@ -9,9 +9,11 @@ import {
   CornerDownRight,
   KeyRound,
   X,
+  AlertCircle,
 } from 'lucide-react';
 import { cn } from '@/lib/utils/cn';
 import type { Source, SynthesisResult, Citation } from '@/modules/ai-search/types';
+import { Button } from '@/components/ui/button';
 
 export interface HistoryItem {
   id: string;
@@ -68,7 +70,9 @@ export function SaiSearchLayout({
   const [historyOpen, setHistoryOpen] = useState(false);
   const [searches, setSearches] = useState<HistoryItem[]>([]);
   const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
   const [cursor, setCursor] = useState<string | null>(null);
+  const cursorRef = useRef<string | null>(null);
   const [hasMore, setHasMore] = useState(false);
   const [pendingDeleteId, setPendingDeleteId] = useState<string | null>(null);
   const pendingDeleteTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
@@ -76,16 +80,24 @@ export function SaiSearchLayout({
 
   const loadHistory = useCallback(async (reset: boolean) => {
     setLoading(true);
+    setError(null);
     try {
       const params = new URLSearchParams({ threaded: '1', limit: '20' });
-      if (!reset && cursor) params.set('cursor', cursor);
+      const currentCursor = reset ? null : cursorRef.current;
+      if (currentCursor) params.set('cursor', currentCursor);
       const res = await fetch(`/api/ai/search-history?${params.toString()}`, {
         headers: { 'Content-Type': 'application/json' },
       });
-      if (!res.ok) return;
+      if (!res.ok) {
+        setError('Failed to load search history. Please try again.');
+        return;
+      }
       const body = await res.json();
       const data = body?.data as { sessions?: Record<string, unknown>[]; nextCursor?: string | null } | undefined;
-      if (!body?.success || !data) return;
+      if (!body?.success || !data) {
+        setError('Failed to load search history. Please try again.');
+        return;
+      }
       const fetched: HistoryItem[] = (data.sessions ?? []).map((s: Record<string, unknown>) => {
         const children = ((s.children as Record<string, unknown>[]) ?? []).map(
           (c: Record<string, unknown>) =>
@@ -101,14 +113,16 @@ export function SaiSearchLayout({
         } as HistoryItem;
       });
       setSearches((prev) => (reset ? fetched : [...prev, ...fetched]));
-      setCursor(data.nextCursor ?? null);
+      const nextCursor = data.nextCursor ?? null;
+      setCursor(nextCursor);
+      cursorRef.current = nextCursor;
       setHasMore(Boolean(data.nextCursor));
     } catch {
-      /* non-fatal */
+      setError('Failed to load search history. Please try again.');
     } finally {
       setLoading(false);
     }
-  }, [cursor]);
+  }, []);
 
   useEffect(() => {
     let cancelled = false;
@@ -118,8 +132,7 @@ export function SaiSearchLayout({
     return () => {
       cancelled = true;
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [loadHistory]);
 
   useEffect(() => {
     return () => {
@@ -138,15 +151,24 @@ export function SaiSearchLayout({
     e.stopPropagation();
     if (pendingDeleteTimer.current) clearTimeout(pendingDeleteTimer.current);
     setPendingDeleteId(null);
+
+    // Optimistic removal
+    setSearches((prev) => {
+      const removed = prev.filter((s) => s.id !== id && s.parentSessionId !== id);
+      return removed;
+    });
+
     try {
       const res = await fetch(`/api/ai/search-history?id=${encodeURIComponent(id)}`, { method: 'DELETE' });
-      if (res.ok) {
-        setSearches((prev) => prev.filter((s) => s.id !== id && s.parentSessionId !== id));
+      if (!res.ok) {
+        // Rollback on failure — reload from server
+        loadHistory(true);
       }
     } catch {
-      /* ignore */
+      // Rollback on failure — reload from server
+      loadHistory(true);
     }
-  }, []);
+  }, [loadHistory]);
 
   const handleDeleteClick = useCallback(
     (id: string, e: React.MouseEvent) => {
@@ -171,14 +193,15 @@ export function SaiSearchLayout({
     return (
       <div key={item.id} className="group relative">
         {isSelected && <span className="absolute left-0 top-1.5 bottom-1.5 w-0.5 rounded-full bg-sai-accent" />}
-        <button type="button"
+        <Button type="button"
           onClick={() => {
             onSelectSession?.(item);
             setHistoryOpen(false);
           }}
           aria-current={isSelected ? 'true' : undefined}
+          variant="ghost"
           className={cn(
-            'w-full text-left pr-3 py-1.5 text-xs rounded-control transition-colors truncate flex items-center gap-1.5',
+            'w-full justify-start pr-3 py-1.5 text-xs rounded-control truncate h-auto',
             isSelected
               ? 'bg-hover/80 text-ink font-medium'
               : 'text-ink-2 hover:text-ink hover:bg-hover/40'
@@ -187,20 +210,22 @@ export function SaiSearchLayout({
         >
           {depth > 0 && <CornerDownRight size={11} className="shrink-0 text-ink-3" />}
           <span className="truncate block flex-1">{label(item)}</span>
-        </button>
+        </Button>
         <span className="absolute right-2 top-1/2 -translate-y-1/2 flex items-center gap-1 opacity-0 group-hover:opacity-100 transition-opacity">
           {!isPendingDelete && <span className="text-xs text-ink-3 pointer-events-none">{item.sourceCount} src</span>}
-          <button type="button"
+          <Button type="button"
             onClick={(e) => handleDeleteClick(item.id, e)}
+            variant="ghost"
+            size="icon-sm"
             className={cn(
-              'p-0.5 transition-colors rounded',
+              'p-0.5',
               isPendingDelete ? 'text-red-400 bg-red-500/10 opacity-100' : 'text-ink-3 hover:text-red-400'
             )}
             aria-label={isPendingDelete ? 'Confirm delete' : 'Remove from history'}
             title={isPendingDelete ? 'Click again to delete' : 'Remove from history'}
           >
             {isPendingDelete ? <Check size={11} /> : <Trash2 size={11} />}
-          </button>
+          </Button>
         </span>
       </div>
     );
@@ -215,30 +240,34 @@ export function SaiSearchLayout({
         {/* Header bar */}
         <div className="h-12 flex items-center justify-between px-4 border-b border-line/60 shrink-0">
           <div className="flex items-center gap-1">
-            <button type="button"
+            <Button type="button"
               onClick={onNewSearch}
-              className="h-8 w-8 flex items-center justify-center rounded-lg text-ink-2 hover:text-ink hover:bg-hover/40 transition-colors"
+              variant="ghost"
+              size="icon"
               title="New search"
             >
               <Plus size={16} />
-            </button>
-            <button type="button"
+            </Button>
+            <Button type="button"
               onClick={() => setHistoryOpen(true)}
-              className="h-8 w-8 flex items-center justify-center rounded-lg text-ink-2 hover:text-ink hover:bg-hover/40 transition-colors"
+              variant="ghost"
+              size="icon"
               title="Search history"
             >
               <Clock size={16} />
-            </button>
+            </Button>
             <span className="ml-2 text-sm font-semibold text-ink tracking-tight">Sai Search</span>
           </div>
-          <button type="button"
+          <Button type="button"
             onClick={onOpenApiKeys}
-            className="h-8 w-8 flex items-center justify-center rounded-lg text-ink-2 hover:text-ink hover:bg-hover/40 transition-colors shrink-0 relative"
+            variant="ghost"
+            size="icon"
             title="API Keys"
+            className="relative"
           >
             <KeyRound size={15} />
             {hasApiKeys && <span className="absolute top-1.5 right-1.5 w-1.5 h-1.5 rounded-full bg-sai-green" />}
-          </button>
+          </Button>
         </div>
 
         <div className="flex-1 overflow-y-auto">{children}</div>
@@ -254,12 +283,13 @@ export function SaiSearchLayout({
           <div className="absolute inset-y-0 left-0 w-80 max-w-[85vw] bg-canvas border-r border-line shadow-overlay flex flex-col animate-in slide-in-from-left duration-200">
             <div className="flex items-center justify-between px-5 pt-5 pb-3 shrink-0">
               <span className="text-sm font-semibold text-ink tracking-tight">History</span>
-              <button type="button"
+              <Button type="button"
                 onClick={() => setHistoryOpen(false)}
-                className="h-7 w-7 flex items-center justify-center rounded-lg text-ink-3 hover:text-ink hover:bg-hover/40 transition-colors"
+                variant="ghost"
+                size="icon-sm"
               >
                 <X size={15} />
-              </button>
+              </Button>
             </div>
 
             <div
@@ -271,7 +301,20 @@ export function SaiSearchLayout({
                 }
               }}
             >
-              {searches.length === 0 ? (
+              {error ? (
+                <div className="flex flex-col items-center justify-center px-5 py-8 text-center">
+                  <AlertCircle size={20} className="text-sai-red mb-2" />
+                  <p className="text-xs text-ink-2 mb-3">{error}</p>
+                  <Button
+                    type="button"
+                    variant="outline"
+                    size="sm"
+                    onClick={() => loadHistory(true)}
+                  >
+                    Retry
+                  </Button>
+                </div>
+              ) : searches.length === 0 ? (
                 <p className="px-5 text-xs text-ink-3 italic">{loading ? 'Loading…' : 'No recent searches'}</p>
               ) : (
                 <div ref={listEndRef} className="px-3">
